baseRecorder.node and baseRecorder.browser
